[nxp][platform][mcxw71_k32w1] Rename mcxw71_k32w1 to mcxw71
 * rename mcxw71_k32w1 to mcxw71
 * remove k32w1, update where needed
