+++ src/platform/nxp/mcxw71/BLEManagerImpl.h
