+++ src/platform/nxp/mcxw71/BUILD.gn
@@ -28,7 +28,7 @@ if (!nxp_external_sdk) {
 }
 
 assert(chip_device_platform == "nxp")
-assert(nxp_platform == "mcxw71_k32w1")
+assert(nxp_platform == "mcxw71")
 
 assert(!nxp_use_low_power || (nxp_use_low_power && !nxp_enable_ot_cli),
        "Please disable low power if openthread CLI is needed!")
@@ -73,7 +73,7 @@ config("nxp_platform_config") {
 
 static_library("nxp_platform") {
   defines = [
-    "NXP_DEVICE_K32W1_MCXW7X=1",
+    "NXP_DEVICE_MCXW7X=1",
     "NXP_USE_MML=1",
   ]
 
@@ -144,8 +144,8 @@ static_library("nxp_platform") {
     "${chip_root}/src/credentials/examples/ExampleDACs.h",
     "${chip_root}/src/credentials/examples/ExamplePAI.h",
     "${chip_root}/src/platform/nxp/common/crypto/PersistentStorageOpKeystoreS200.h",
-    "${chip_root}/src/platform/nxp/mcxw71_k32w1/BLEManagerImpl.h",
-    "${chip_root}/src/platform/nxp/mcxw71_k32w1/SMU2Manager.h",
+    "${chip_root}/src/platform/nxp/mcxw71/BLEManagerImpl.h",
+    "${chip_root}/src/platform/nxp/mcxw71/SMU2Manager.h",
   ]
 
   public_deps = [ "${chip_root}/src/platform:platform_base" ]
