+++ src/platform/nxp/mcxw71/BlePlatformConfig.h
