+++ src/platform/nxp/mcxw71/ConfigurationManagerImpl.h
@@ -26,7 +26,7 @@
 #pragma once
 
 #if (CHIP_PLAT_NVM_SUPPORT == 1)
-#include <platform/nxp/mcxw71_k32w1/K32W1Config.h>
+#include <platform/nxp/mcxw71/K32W1Config.h>
 #else
 #include <platform/nxp/common/NXPConfig.h>
 #endif
