+++ src/platform/nxp/mcxw71/ConnectivityManagerImpl.cpp
