+++ src/platform/nxp/mcxw71/LowPowerHooks.cpp
