+++ src/platform/nxp/mcxw71/PlatformManagerImpl.cpp
@@ -33,10 +33,10 @@
 #ifdef EXTERNAL_CONFIGURATIONMANAGERIMPL_HEADER
 #include EXTERNAL_CONFIGURATIONMANAGERIMPL_HEADER
 #else
-#include <platform/nxp/mcxw71_k32w1/ConfigurationManagerImpl.h>
+#include <platform/nxp/mcxw71/ConfigurationManagerImpl.h>
 #endif
-#include <platform/nxp/mcxw71_k32w1/DiagnosticDataProviderImpl.h>
-#include <platform/nxp/mcxw71_k32w1/SMU2Manager.h>
+#include <platform/nxp/mcxw71/DiagnosticDataProviderImpl.h>
+#include <platform/nxp/mcxw71/SMU2Manager.h>
 
 #if CHIP_SYSTEM_CONFIG_USE_LWIP
 #include <lwip/tcpip.h>
