+++ src/platform/nxp/mcxw71/PlatformManagerImpl.h
