+++ src/platform/nxp/mcxw71/args.gni
@@ -18,8 +18,8 @@ import("//build_overrides/openthread.gni")
 
 openthread_root =
     "//third_party/connectedhomeip/third_party/openthread/ot-nxp/openthread"
-nxp_platform = "mcxw71_k32w1"
-nxp_sdk_name = "mcxw71_k32w1_sdk"
+nxp_platform = "mcxw71"
+nxp_sdk_name = "mcxw7x_sdk"
 nxp_device_layer = "nxp/${nxp_platform}"
 nxp_nvm_component = "nvs"
 nxp_use_lwip = false
@@ -56,6 +56,6 @@ openthread_external_mbedtls = mbedtls_target
 openthread_project_core_config_file = "OpenThreadConfig.h"
 openthread_core_config_platform_check_file =
     "openthread-core-mcxw71-config-check.h"
-openthread_core_config_deps = [ "${chip_root}/third_party/openthread/platforms/nxp/${nxp_platform}:openthread_core_config_k32w1" ]
+openthread_core_config_deps = [ "${chip_root}/third_party/openthread/platforms/nxp/${nxp_platform}:openthread_core_config_mcxw71" ]
 
-openthread_external_platform = "${chip_root}/third_party/openthread/platforms/nxp/${nxp_platform}:libopenthread-k32w1"
+openthread_external_platform = "${chip_root}/third_party/openthread/platforms/nxp/${nxp_platform}:libopenthread-mcxw71"
