+++ src/platform/nxp/mcxw71/ram_storage.c
